added composable components (Map, MapSlider)
