+++ app/src/main/java/com/kodeco/android/coordplot/components/Map.kt
@@ -1,0 +1,38 @@
+package com.kodeco.android.coordplot.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.paint
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.kodeco.android.coordplot.R
+import com.kodeco.android.coordplot.ui.theme.MyApplicationTheme
+
+@Composable
+fun Map(xPercent: Float, yPercent: Float, modifier: Modifier = Modifier) {
+    Box(
+        modifier = Modifier
+            .size(300.dp)
+            .background(color = Color.Blue)
+    ) {
+        Box(
+            modifier = Modifier
+                .offset((xPercent * 300 - 18).dp, (yPercent * 300 - 18).dp)
+                .paint(painter = painterResource(id = R.drawable.star))
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun MapPreview() {
+    MyApplicationTheme {
+        Map(xPercent = 0.5f, yPercent = 0.5f)
+    }
+}
